+++ cascade-api/src/Cascade/Api/Database/Task.hs
@@ -6,9 +6,6 @@ module Cascade.Api.Database.Task
 
 import qualified Cascade.Api.Data.Task         as Task
 import qualified Cascade.Api.Data.WrappedC     as Wrapped
-import           Control.Lens                   ( _Wrapped'
-                                                , view
-                                                )
 import           Data.Generics.Labels           ( )
 import           Database.Beam                  ( Beamable
                                                 , C
